Added score, max score is set at 9999999.
If you want a higher max score it needs to be adjusted accordingly with
max digits or it will try to create an array with negative entries and
crash.

Hacked a printRight method from printCentred method. There's an
unnecessary width variable at the moment, but the function works fine
with it set to 0 so I've left it. Feel free to clean it up if you
understand how.